Refactoring packages and fixed constructor of ActivationElement
