+++ src/main/java/dk/dtu/compute/se/pisd/roborally/controller/boardElementController/BoardLaserController.java
@@ -5,8 +5,7 @@ import dk.dtu.compute.se.pisd.roborally.model.Command;
 import dk.dtu.compute.se.pisd.roborally.model.CommandCard;
 import dk.dtu.compute.se.pisd.roborally.model.Player;
 import dk.dtu.compute.se.pisd.roborally.model.Space;
-import dk.dtu.compute.se.pisd.roborally.model.boardElement.BoardLaser;
-import dk.dtu.compute.se.pisd.roborally.model.boardElement.MoveHazard;
+import dk.dtu.compute.se.pisd.roborally.model.boardElement.activationElements.BoardLaser;
 
 public class BoardLaserController implements IBoardElementController, ILaser {
     private BoardLaser model;
